pass[] data type optimization

+++ radio/src/pulses/pxx_arm.cpp
@@ -216,7 +216,7 @@ void setupPulsesPXX(unsigned int port)
   putPcmByte(0, port);
 
   /* PPM */
-  static uint32_t pass[NUM_MODULES] = { MODULES_INIT(0) };
+  static uint8_t pass[NUM_MODULES] = { MODULES_INIT(0) };
   int sendUpperChannels = 0;
   if (pass[port]++ & 0x01) {
     sendUpperChannels = g_model.moduleData[port].channelsCount;
